ABOUT
Cars is a mobile application which runs on any smart phone with Android
platform. Android is a software platform and operating system for
mobile devices based on the Linux operating system and developed by
Google and the Open Handset Alliance. It allows developers to write
managed code in a Java-like language that utilizes Google-developed
Java libraries, but does not support programs developed in native code.
It uses XML for the User Interface. The Java and XML can be bound
together to build the applications
 Cars is an informative application which allows the user to select a
car based on his choice of Make, Class and Price-Range. The information
includes various key components like the Fuel consumption,
dimension,power and so on which help the user to select the car of his
choice. Cars is targeted for the latest version of the Android
operating system Android 5.0 "Lollipop" (API level 21). It also has the
capability to support the lower versions until Android 3.2
"Honeycomb"(API level 13)